spy_relayer: move Relayer.runAuditor()
Make it part of the backend since the backend has all of the payload
specific logic into it.
