+++ relayer/spy_relayer/src/backends/definitions.ts
@@ -38,7 +38,7 @@ export interface Relayer {
   /** Process the request to relay a message */
   process(key: string, privKey: any, logger: ScopedLogger): Promise<void>;
 
-  /** Run and auditor to ensure the relay was not rolled back due to a chain reorg */
+  /** Run an auditor to ensure the relay was not rolled back due to a chain reorg */
   runAuditor(workerInfo: WorkerInfo): Promise<void>;
 }
